refactor: common routines between sha algorithms

+++ wks/hash_functions/rs_sha1/src/sha1state.rs
@@ -5,7 +5,7 @@ use core::{
     ops::AddAssign,
 };
 use hash_ctx_lib::{BlockHasher, GenericStateHasher, HasherWords};
-use internal_state::{LOWER_HEX_ERR, Sha160BitsState, UPPER_HEX_ERR};
+use internal_state::{Sha160BitsState, LOWER_HEX_ERR, UPPER_HEX_ERR};
 
 pub(crate) const H0: u32 = 0x67452301;
 pub(crate) const H1: u32 = 0xEFCDAB89;
@@ -63,7 +63,7 @@ impl BuildHasher for Sha1State {
         Sha1Hasher {
             size: u64::default(),
             state: self.clone(),
-            padding: [0u8; Sha1Hasher::U8_PAD_SIZE],
+            padding: [0u8; Sha1Hasher::U8_PAD_SIZE as usize],
         }
     }
 }
